fix for fee logging

+++ programs/m4a_protocol/src/lib.rs
@@ -197,7 +197,7 @@ fn apply_fee<'info>(
     //Transfer fee to Treasurer Wallet
     token::transfer(cpi_ctx, fixed_pointed_notation_amount)?;
     
-    msg!("Successfully transferred ${:.2} as fee to: {}", (amount/100) as f32, treasurer.address);
+    msg!("Successfully transferred ${:.2} as fee to: {}", amount as f64 / 100.0, treasurer.address);
 
     Ok(())
 }
